Fixed an issue where filter action execution was slow when modifying more than a few events at a time, as is common when manually executing filters on a large folder.  Most DB transactions are now deferred until the end of the filtering operation, allowing them to be done in bulk (many fewer overall transactions). Added a button to the Filter List to "Run enabled filters against all events". The folder context menu now offers the ability to run all enabled filters. Misc tweaks.

+++ ErrorTrackerServer/Controllers/EventData.cs
@@ -199,7 +199,7 @@ namespace ErrorTrackerServer.Controllers
 	}
 	public class MoveEventRequest : GetEventRequest
 	{
-		public long newFolderId;
+		public int newFolderId;
 	}
 	public class GetEventsRequest : ProjectRequestBase
 	{
@@ -207,7 +207,7 @@ namespace ErrorTrackerServer.Controllers
 	}
 	public class MoveEventsRequest : GetEventsRequest
 	{
-		public long newFolderId;
+		public int newFolderId;
 	}
 	public class SetEventsColorRequest : GetEventsRequest
 	{
